# openapi.yaml
openapi: 3.1.0
info:
  title: assessment-service
  version: v0.3.1
paths:
...
  /v1/admin/content/import:
    post:
      tags:
      - assessment
      - assessment
      summary: Import Content
      operationId: postAdminContentImport
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentImportRequest'
            example:
              source_id: string
              contract_version: string
              validate_only: true
              error_mode: collect
              payload:
                subjects: []
                topics: []
                micro_skills: []
                tests: []
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentImportResponse'
              example:
                import_id: string
                source_id: string
                imported: 1
                status: string
                errors:
                - code: string
                  message: string
                  path: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - bearerAuth: []
...
  /v1/admin/subjects:
    get:
      tags:
      - assessment
      - assessment
      summary: List Subjects
      operationId: getAdminSubjects
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/SubjectResponse'
                type: array
                title: Response List Subjects V1 Admin Subjects Get
              example:
              - code: string
                name: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - bearerAuth: []
    post:
      tags:
      - assessment
      - assessment
      summary: Create Subject
      operationId: postAdminSubjects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectCreateRequest'
            example:
              code: string
              name: string
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectResponse'
              example:
                code: string
                name: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - bearerAuth: []
  /v1/admin/topics:
    get:
      tags:
      - assessment
      - assessment
      summary: List Topics
      operationId: getAdminTopics
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TopicResponse'
                type: array
                title: Response List Topics V1 Admin Topics Get
              example:
              - code: string
                subject_code: string
                grade: 1
                name: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - bearerAuth: []
    post:
      tags:
      - assessment
      - assessment
      summary: Create Topic
      operationId: postAdminTopics
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TopicCreateRequest'
            example:
              code: string
              subject_code: string
              grade: 1
              name: string
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicResponse'
              example:
                code: string
                subject_code: string
                grade: 1
                name: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - bearerAuth: []
  /v1/admin/micro-skills:
    get:
      tags:
      - assessment
      - assessment
      summary: List Micro Skills
      operationId: getAdminMicroSkills
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/MicroSkillResponse'
                type: array
                title: Response List Micro Skills V1 Admin Micro Skills Get
              example:
              - node_id: string
                subject_code: string
                topic_code: string
                grade: 1
                section_code: string
                section_name: string
                micro_skill_name: string
                predecessor_ids:
                - string
                criticality: low
                source_ref: string
                description: string
                status: draft
                external_ref: string
                version: 1
                created_at: '2026-03-13T00:00:00Z'
                updated_at: '2026-03-13T00:00:00Z'
                blocks_count: 1
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - bearerAuth: []
    post:
      tags:
      - assessment
      - assessment
      summary: Create Micro Skill
      operationId: postAdminMicroSkills
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MicroSkillCreateRequest'
            example:
              node_id: string
              subject_code: string
              topic_code: string
              grade: 1
              section_code: string
              section_name: string
              micro_skill_name: string
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MicroSkillResponse'
              example:
                node_id: string
                subject_code: string
                topic_code: string
                grade: 1
                section_code: string
                section_name: string
                micro_skill_name: string
                predecessor_ids:
                - string
                criticality: low
                source_ref: string
                description: string
                status: draft
                external_ref: string
                version: 1
                created_at: '2026-03-13T00:00:00Z'
                updated_at: '2026-03-13T00:00:00Z'
                blocks_count: 1
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - bearerAuth: []
  /v1/admin/micro-skills/{node_id}:
    put:
      tags:
      - assessment
      - assessment
      summary: Update Micro Skill
      operationId: putAdminMicroSkillsByNodeId
      parameters:
      - name: node_id
        in: path
        required: true
        schema:
          type: string
          title: Node Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MicroSkillUpdateRequest'
            example:
              subject_code: string
              topic_code: string
              grade: 1
              section_code: string
              section_name: string
              micro_skill_name: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MicroSkillResponse'
              example:
                node_id: string
                subject_code: string
                topic_code: string
                grade: 1
                section_code: string
                section_name: string
                micro_skill_name: string
                predecessor_ids:
                - string
                criticality: low
                source_ref: string
                description: string
                status: draft
                external_ref: string
                version: 1
                created_at: '2026-03-13T00:00:00Z'
                updated_at: '2026-03-13T00:00:00Z'
                blocks_count: 1
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - bearerAuth: []
    delete:
      tags:
      - assessment
      - assessment
      summary: Delete Micro Skill
      operationId: deleteAdminMicroSkillsByNodeId
      parameters:
      - name: node_id
        in: path
        required: true
        schema:
          type: string
          title: Node Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - bearerAuth: []
  /v1/admin/micro-skills/{node_id}/links:
    post:
      tags:
      - assessment
      - assessment
      summary: Link Micro Skill Predecessors
      operationId: postAdminMicroSkillsLinksByNodeId
      parameters:
      - name: node_id
        in: path
        required: true
        schema:
          type: string
          title: Node Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MicroSkillLinkRequest'
            example:
              predecessor_ids:
              - string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MicroSkillResponse'
              example:
                node_id: string
                subject_code: string
                topic_code: string
                grade: 1
                section_code: string
                section_name: string
                micro_skill_name: string
                predecessor_ids:
                - string
                criticality: low
                source_ref: string
                description: string
                status: draft
                external_ref: string
                version: 1
                created_at: '2026-03-13T00:00:00Z'
                updated_at: '2026-03-13T00:00:00Z'
                blocks_count: 1
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    ContentImportDetails:
      properties:
        subjects_created:
          type: integer
          title: Subjects Created
          default: 0
        subjects_updated:
          type: integer
          title: Subjects Updated
          default: 0
        topics_created:
          type: integer
          title: Topics Created
          default: 0
        topics_updated:
          type: integer
          title: Topics Updated
          default: 0
        micro_skills_created:
          type: integer
          title: Micro Skills Created
          default: 0
        micro_skills_updated:
          type: integer
          title: Micro Skills Updated
          default: 0
        tests_created:
          type: integer
          title: Tests Created
          default: 0
        tests_updated:
          type: integer
          title: Tests Updated
          default: 0
        tests_failed:
          type: integer
          title: Tests Failed
          default: 0
        questions_failed:
          type: integer
          title: Questions Failed
          default: 0
      type: object
      title: ContentImportDetails
    ContentImportIssue:
      properties:
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
        path:
          type: string
          title: Path
      type: object
      required:
      - code
      - message
      - path
      title: ContentImportIssue
...
    ContentImportPayload:
      properties:
        subjects:
          items:
            $ref: '#/components/schemas/ContentImportSubjectItem'
          type: array
          title: Subjects
        topics:
          items:
            $ref: '#/components/schemas/ContentImportTopicItem'
          type: array
          title: Topics
        micro_skills:
          items:
            $ref: '#/components/schemas/ContentImportMicroSkillItem'
          type: array
          title: Micro Skills
        tests:
          items:
            $ref: '#/components/schemas/ContentImportTestItem'
          type: array
          title: Tests
      type: object
      title: ContentImportPayload
...
    ContentImportRequest:
      properties:
        source_id:
          type: string
          title: Source Id
        contract_version:
          type: string
          title: Contract Version
        validate_only:
          type: boolean
          title: Validate Only
          default: false
        error_mode:
          type: string
          enum:
          - collect
          - fail_fast
          title: Error Mode
          default: collect
        payload:
          $ref: '#/components/schemas/ContentImportPayload'
      type: object
      required:
      - source_id
      - contract_version
      - payload
      title: ContentImportRequest
    ContentImportResponse:
      properties:
        import_id:
          type: string
          title: Import Id
        source_id:
          type: string
          title: Source Id
        imported:
          type: integer
          title: Imported
        status:
          type: string
          title: Status
        errors:
          items:
            $ref: '#/components/schemas/ContentImportIssue'
          type: array
          title: Errors
        warnings:
          items:
            $ref: '#/components/schemas/ContentImportIssue'
          type: array
          title: Warnings
        details:
          anyOf:
          - $ref: '#/components/schemas/ContentImportDetails'
          - type: 'null'
      type: object
      required:
      - import_id
      - source_id
      - imported
      - status
      title: ContentImportResponse
...
    CriticalityLevel:
      type: string
      enum:
      - low
      - medium
      - high
      title: CriticalityLevel
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    MicroSkillCreateRequest:
      properties:
        node_id:
          type: string
          title: Node Id
        subject_code:
          type: string
          title: Subject Code
        topic_code:
          type: string
          title: Topic Code
        grade:
          type: integer
          maximum: 11.0
          minimum: 1.0
          title: Grade
        section_code:
          type: string
          title: Section Code
        section_name:
          type: string
          title: Section Name
        micro_skill_name:
          type: string
          title: Micro Skill Name
        predecessor_ids:
          items:
            type: string
          type: array
          title: Predecessor Ids
        criticality:
          $ref: '#/components/schemas/CriticalityLevel'
          default: medium
        source_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Ref
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        status:
          $ref: '#/components/schemas/MicroSkillStatus'
          default: active
        external_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: External Ref
      type: object
      required:
      - node_id
      - subject_code
      - topic_code
      - grade
      - section_code
      - section_name
      - micro_skill_name
      title: MicroSkillCreateRequest
    MicroSkillLinkRequest:
      properties:
        predecessor_ids:
          items:
            type: string
          type: array
          title: Predecessor Ids
      type: object
      required:
      - predecessor_ids
      title: MicroSkillLinkRequest
    MicroSkillResponse:
      properties:
        node_id:
          type: string
          title: Node Id
        subject_code:
          type: string
          title: Subject Code
        topic_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Topic Code
        grade:
          type: integer
          title: Grade
        section_code:
          type: string
          title: Section Code
        section_name:
          type: string
          title: Section Name
        micro_skill_name:
          type: string
          title: Micro Skill Name
        predecessor_ids:
          items:
            type: string
          type: array
          title: Predecessor Ids
        criticality:
          $ref: '#/components/schemas/CriticalityLevel'
        source_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Ref
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        status:
          $ref: '#/components/schemas/MicroSkillStatus'
        external_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: External Ref
        version:
          type: integer
          title: Version
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        blocks_count:
          type: integer
          title: Blocks Count
      type: object
      required:
      - node_id
      - subject_code
      - topic_code
      - grade
      - section_code
      - section_name
      - micro_skill_name
      - predecessor_ids
      - criticality
      - source_ref
      - description
      - status
      - external_ref
      - version
      - created_at
      - updated_at
      - blocks_count
      title: MicroSkillResponse
    MicroSkillStatus:
      type: string
      enum:
      - draft
      - active
      - archived
      title: MicroSkillStatus
    MicroSkillUpdateRequest:
      properties:
        subject_code:
          type: string
          title: Subject Code
        topic_code:
          type: string
          title: Topic Code
        grade:
          type: integer
          maximum: 11.0
          minimum: 1.0
          title: Grade
        section_code:
          type: string
          title: Section Code
        section_name:
          type: string
          title: Section Name
        micro_skill_name:
          type: string
          title: Micro Skill Name
        predecessor_ids:
          items:
            type: string
          type: array
          title: Predecessor Ids
        criticality:
          $ref: '#/components/schemas/CriticalityLevel'
          default: medium
        source_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Ref
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        status:
          $ref: '#/components/schemas/MicroSkillStatus'
          default: active
        external_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: External Ref
      type: object
      required:
      - subject_code
      - topic_code
      - grade
      - section_code
      - section_name
      - micro_skill_name
      title: MicroSkillUpdateRequest
...
    SubjectCreateRequest:
      properties:
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
      type: object
      required:
      - code
      - name
      title: SubjectCreateRequest
    SubjectResponse:
      properties:
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
      type: object
      required:
      - code
      - name
      title: SubjectResponse
...
    TopicCreateRequest:
      properties:
        code:
          type: string
          title: Code
        subject_code:
          type: string
          title: Subject Code
        grade:
          type: integer
          maximum: 11.0
          minimum: 1.0
          title: Grade
        name:
          type: string
          title: Name
      type: object
      required:
      - code
      - subject_code
      - grade
      - name
      title: TopicCreateRequest
    TopicResponse:
      properties:
        code:
          type: string
          title: Code
        subject_code:
          type: string
          title: Subject Code
        grade:
          type: integer
          title: Grade
        name:
          type: string
          title: Name
      type: object
      required:
      - code
      - subject_code
      - grade
      - name
      title: TopicResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://example.com/problems/authentication
            title: Unauthorized
            status: 401
            detail: Bearer access token is required
            code: UNAUTHORIZED
            request_id: 00000000-0000-4000-8000-000000000000
    Forbidden:
      description: Forbidden
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://example.com/problems/access-denied
            title: Forbidden
            status: 403
            detail: Access denied
            code: FORBIDDEN
            request_id: 00000000-0000-4000-8000-000000000000
    TooManyRequests:
      description: Too Many Requests
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://example.com/problems/too-many-requests
            title: Too Many Requests
            status: 429
            detail: Rate limit exceeded
            code: RATE_LIMITED
            request_id: 00000000-0000-4000-8000-000000000000
servers:
- url: http://localhost:8003
  description: Local
- url: https://api.monitoring.example
  description: Production
security:
- bearerAuth: []